@@ -11,48 +11,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~11.0.5",
-    "@angular/forms": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
-    "angular-font-awesome": "^3.1.2",
-    "bootstrap": "^4.3.1",
-    "core-js": "^2.5.4",
-    "express": "^4.19.2",
-    "file-saver": "^2.0.2",
-    "font-awesome": "^4.7.0",
-    "moment": "^2.29.4",
-    "ngx-bootstrap": "^5.1.1",
-    "rxjs": "~6.3.3",
-    "tslib": "^1.9.0",
-    "xlsx": "^0.17.0",
-    "zone.js": "~0.8.26"
+    "@angular/animations": "~15.0.0",
+    "@angular/common": "~15.0.0",
+    "@angular/compiler": "~15.0.0",
+    "@angular/core": "~15.0.0",
+    "@angular/forms": "~15.0.0",
+    "@angular/platform-browser": "~15.0.0",
+    "@angular/platform-browser-dynamic": "~15.0.0",
+    "@angular/router": "~15.0.0",
+    "bootstrap": "^5.2.0",
+    "core-js": "^3.25.0",
+    "file-saver": "^2.0.5",
+    "rxjs": "~7.5.0",
+    "tslib": "^2.4.0",
+    "xlsx": "^0.18.5",
+    "zone.js": "~0.12.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^15.0.0",
-    "@angular/cli": "~15.0.0",
-    "@angular/compiler-cli": "~18.0.2",
-    "@angular/language-service": "~7.2.0",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "handlebars": ">=4.7.7",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~6.3.16",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.4",
-    "serialize-javascript": ">=3.1.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "@angular/cli": "^15.2.11",
+    "@angular/compiler-cli": "~15.0.0",
+    "@angular/language-service": "~15.0.0",
+    "@types/jasmine": "~4.3.0",
+    "@types/node": "^18.7.18",
+    "jasmine-core": "~4.4.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "typescript": "~4.8.0"
   }
 }
