@@ -1,20 +1,137 @@
+require('dotenv').config();
 const express = require('express');
-//const cors = require('cors');
+const cors = require('cors');
+const helmet = require('helmet');
 const mongoose = require('mongoose');
 const requireDir = require('require-dir');
-var bodyParser = require('body-parser');
+const logger = require('./src/utils/logger');
 
 const app = express();
 
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
+// Trust proxy for rate limiting behind reverse proxy
+app.set('trust proxy', 1);
 
-//app.use(express.json());
-//app.use(cors);
+// Security middleware
+app.use(helmet({
+  crossOriginEmbedderPolicy: false,
+  contentSecurityPolicy: {
+    directives: {
+      defaultSrc: ["'self'"],
+      styleSrc: ["'self'", "'unsafe-inline'"],
+      scriptSrc: ["'self'"],
+      imgSrc: ["'self'", "data:", "https:"],
+    },
+  },
+}));
 
-mongoose.connect('mongodb://localhost:27017/nodeapi', { useNewUrlParser: true });
+// CORS configuration
+const corsOptions = {
+  origin: process.env.CORS_ORIGIN ? process.env.CORS_ORIGIN.split(',') : 'http://localhost:4200',
+  methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
+  allowedHeaders: ['Content-Type', 'Authorization', 'X-Requested-With'],
+  credentials: true,
+  maxAge: 86400 // 24 hours
+};
+app.use(cors(corsOptions));
+
+// Body parser middleware with size limits
+app.use(express.json({
+  limit: '1mb',
+  strict: true
+}));
+app.use(express.urlencoded({
+  extended: true,
+  limit: '1mb'
+}));
+
+// Request logging middleware
+app.use((req, res, next) => {
+  logger.http(`${req.method} ${req.originalUrl} - ${req.ip}`);
+  next();
+});
+
+// Database connection with better error handling
+const mongoUri = process.env.MONGODB_URI || 'mongodb://localhost:27017/nodeapi';
+
+mongoose.connection.on('connected', () => {
+  logger.info('MongoDB connected successfully');
+});
+
+mongoose.connection.on('error', (err) => {
+  logger.error('MongoDB connection error:', err);
+});
+
+mongoose.connection.on('disconnected', () => {
+  logger.warn('MongoDB disconnected');
+});
+
+// Graceful shutdown
+process.on('SIGINT', async () => {
+  logger.info('Shutting down gracefully...');
+  await mongoose.connection.close();
+  process.exit(0);
+});
+
+mongoose.connect(mongoUri, {
+  useNewUrlParser: true,
+  useUnifiedTopology: true,
+})
+.then(() => {
+  logger.info('Database connection established');
+})
+.catch(err => {
+  logger.error('MongoDB connection error:', err);
+  process.exit(1);
+});
+
+// Load models
 requireDir('./src/models');
 
+// Health check endpoint (before rate limiting)
+app.get('/health', (req, res) => {
+  res.json({
+    status: 'OK',
+    timestamp: new Date().toISOString(),
+    uptime: process.uptime(),
+    environment: process.env.NODE_ENV || 'development',
+    version: process.env.npm_package_version || '1.0.0'
+  });
+});
+
+// API Routes
 app.use('/api', require('./src/routes'));
 
-app.listen(3001);
+// Global error handling middleware
+app.use((err, req, res, next) => {
+  logger.error(`Error: ${err.message}`, {
+    stack: err.stack,
+    url: req.originalUrl,
+    method: req.method,
+    ip: req.ip
+  });
+
+  // Don't leak error details in production
+  const isDevelopment = process.env.NODE_ENV === 'development';
+
+  res.status(err.status || 500).json({
+    error: isDevelopment ? err.message : 'Internal server error',
+    ...(isDevelopment && { stack: err.stack })
+  });
+});
+
+// 404 handler
+app.use('*', (req, res) => {
+  logger.warn(`404 - Route not found: ${req.method} ${req.originalUrl} - ${req.ip}`);
+  res.status(404).json({
+    error: 'Route not found',
+    path: req.originalUrl,
+    method: req.method
+  });
+});
+
+const port = process.env.PORT || 3001;
+app.listen(port, () => {
+  logger.info(`Server running on port ${port}`);
+  logger.info(`Environment: ${process.env.NODE_ENV || 'development'}`);
+  logger.info(`CORS Origin: ${process.env.CORS_ORIGIN || 'http://localhost:4200'}`);
+});
